# Optimise log redaction tagging