Ahead of enabling tagging by default - this is an easy win for speeding it up (clog does `fmt.Sprintf` under the hood, as well as `interface{}` stuff which has the potential to allocate.

```
$ go1.13.4 test -run=- -bench=^BenchmarkUserDataRedact$ -benchmem -count=10 ./base > new.txt
$ git checkout master
$ go1.13.4 test -run=- -bench=^BenchmarkUserDataRedact$ -benchmem -count=10 ./base > old.txt
$ benchstat -alpha 1 old.txt new.txt
name                            old time/op    new time/op    delta
UserDataRedact/Enabled-12          204ns ± 1%      22ns ± 2%   -89.25%  (p=0.000 n=10+10)
UserDataRedact/EnabledSlice-12    1.28µs ± 1%    0.73µs ± 1%   -42.94%  (p=0.000 n=9+10)
UserDataRedact/Disabled-12        2.18ns ± 1%    0.49ns ± 3%   -77.73%  (p=0.000 n=10+10)

name                            old alloc/op   new alloc/op   delta
UserDataRedact/Enabled-12          64.0B ± 0%      0.0B       -100.00%  (p=0.000 n=10+10)
UserDataRedact/EnabledSlice-12      624B ± 0%      480B ± 0%   -23.08%  (p=0.000 n=10+10)
UserDataRedact/Disabled-12         0.00B          0.00B           ~     (all equal)

name                            old allocs/op  new allocs/op  delta
UserDataRedact/Enabled-12           4.00 ± 0%      0.00       -100.00%  (p=0.000 n=10+10)
UserDataRedact/EnabledSlice-12      27.0 ± 0%      18.0 ± 0%   -33.33%  (p=0.000 n=10+10)
UserDataRedact/Disabled-12          0.00           0.00           ~     (all equal)
```